Added gradle function to generate header
Changed all variables and returns to be bytes as we don't need int for
that.

+++ RemoteControlAPI/src/main/java/com/github/unafraid/remote/control/api/RCDeviceType.java
@@ -37,7 +37,7 @@ public enum RCDeviceType
 		_type = (byte) type;
 	}
 	
-	public int getValue()
+	public byte getValue()
 	{
 		return _type;
 	}
